dailyScrapers/publicoNewsScraper_test.go
package dailyScrapers

import (
	"fmt"
	"github.com/stretchr/testify/assert"
	"news-scrapers-workers-go/models"
	"testing"
	"time"

	"github.com/joho/godotenv"
)

func TestScraperPublico(t *testing.T) {
	e := godotenv.Load()
	if e != nil {
		fmt.Print(e)
	}
	config := models.ScrapingConfig{ScraperId: "testScraperPublico", DeviceID: "testDevicePublico"}
	scraper := PublicoNewsScraper{Config: config}
	baseUrl := "https://www.publico.es/sociedad/sindicatos-medicos-denuncian-sanidad-le-exigen-suministre-material-proteccion.html"
	newUrl := models.UrlNew{baseUrl, time.Now()}

	result := scraper.ScrapNewUrl(newUrl)
	assert.NotEqual(t, result.Content, "")
	assert.NotEmpty(t, result.Tags)
	assert.NotEqual(t, result.Date.IsZero(),true)
	assert.NotEqual(t, result.Headline, "")

}
